Fix cppcheck warnings and modernize C++ practices
- Replace C-style casts with reinterpret_cast/static_cast/const_cast
- Add missing member initializers to constructors (uninitMemberVar)
- Replace NULL/0 with nullptr in pointer contexts
- Add missing override specifiers to virtual methods
- Mark final classes with final keyword
- Fix shadowed variables and signed/unsigned comparisons
- Use pre-increment (++it) over post-increment where applicable
- Refactor Options.cpp to use string_view and StrCaseCmp helpers
- Fix QueueScriptCoordinator::CheckQueue selecting last item instead
  of highest-priority event
- Fix DiskState::LoadNzbInfo losing lsFailed/lsRetry URL statuses
  when restoring queue state from disk
- Remove ServerJoinGroupValidator: m_joinGroup is bool, making the
  0-99 range check trivially unreachable
- Update CI: use modern cmake -S/-B syntax, fix compile_commands.json
  path, restrict cppcheck to warning/performance/portability checks
- Add .clang-tidy and .clangd project configuration
- Update cppcheck invocation in docs/POSIX.md

+++ daemon/frontend/LoggableFrontend.cpp
@@ -56,7 +56,7 @@ void LoggableFrontend::Update()
 		if (!messages->empty())
 		{
 			Message& firstMessage = messages->front();
-			int start = m_neededLogFirstId - firstMessage.GetId() + 1;
+			int start = static_cast<int>(static_cast<int64>(m_neededLogFirstId) - static_cast<int64>(firstMessage.GetId()) + 1);
 			if (start < 0)
 			{
 				PrintSkip();
